requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-classic
streamlit
pypdf
plotly
python-dotenv
chromadb
langchain-chroma
